MADlib image loader:  load dataset from disk
Adds to the madlib_image_loader module ability to load images
directly from disk (instead of having to be in numpy arrays).

New functionality can either be called as a module or run from
the command line.  (Loading from a numpy array still only
supported when called as a module.)

Dataset to be loaded is expected to be organized by image
category/class.  Each directory name labels a specific class,
and only images from that class should be in that directory.

An additional img_name column is added to the table when used in
this way, to help with tracking which image is which later.

Madlib Image Loader Demo notebook has been updated to include a
minor interface change (original API endpoint renamed to
load_dataset_from_np) and an example of loading from disk with
newly added endpoint load_dataset_from_disk.

Closes #16

Co-authored-by:  Nikhil Kak <nkak@pivotal.io>